tests: split json checker out of observer
JsonTestCheck is used widely by our unit tests, not just in pkg/observer. So let's split
this functionality out of the observer to make things more modular.
