+++ pkg/sensors/tracing/kprobe_filterchange_test.go
@@ -16,6 +16,7 @@ import (
 	bc "github.com/cilium/tetragon/api/v1/tetragon/codegen/eventchecker/matchers/bytesmatcher"
 	lc "github.com/cilium/tetragon/api/v1/tetragon/codegen/eventchecker/matchers/listmatcher"
 	sm "github.com/cilium/tetragon/api/v1/tetragon/codegen/eventchecker/matchers/stringmatcher"
+	"github.com/cilium/tetragon/pkg/jsonchecker"
 	"github.com/cilium/tetragon/pkg/kernels"
 	"github.com/cilium/tetragon/pkg/observer"
 	"github.com/cilium/tetragon/pkg/testutils"
@@ -99,7 +100,7 @@ func TestKprobeNSChanges(t *testing.T) {
 	checker := ec.NewUnorderedEventChecker(
 		kprobeChecker,
 	)
-	err = observer.JsonTestCheck(t, checker)
+	err = jsonchecker.JsonTestCheck(t, checker)
 	assert.NoError(t, err)
 }
 
@@ -182,7 +183,7 @@ func testKprobeCapChanges(t *testing.T, spec string, op string, value string) {
 		kprobeChecker,
 	)
 
-	err = observer.JsonTestCheck(t, checker)
+	err = jsonchecker.JsonTestCheck(t, checker)
 	assert.NoError(t, err)
 }
